Split `StringSortComparator` into ICU and ICU-free variants
Refactor the `LocaleManager` and the string comparators so that an
ICU-free (bytewise) variant can be selected at compile time, as a
preparatory step towards an optional ICU dependency.

- Add `LocaleManager.h`, containing the shared base `LocaleManagerBase`
  (with `SortKey`/`SortKeyView` and the `getPrefixSortKeyImpl` helper),
  the ICU-based `LocaleManagerICU` (moved out of `StringSortComparator.h`),
  the bytewise `LocaleManagerNoICU`, and the `LocaleManager` typedef that
  selects between them via the `QLEVER_NO_UNICODE` macro. As a
  preparatory step `LocaleManagerNoICU` still reuses a few ICU-based
  helpers (e.g. `ad_utility::utf8ToLower`).
- `StringSortComparator.h` now only holds the generic comparator
  templates `SimpleStringComparatorImpl` / `TripleComponentComparatorImpl`
  (previously the concrete `SimpleStringComparator` /
  `TripleComponentComparator`), their ICU and NoICU instantiations, and
  the `SimpleStringComparator` / `TripleComponentComparator` typedef
  switch.
- Add the `NO_UNICODE` CMake option. As a preparatory step it only
  defines the `QLEVER_NO_UNICODE` macro (switching the typedefs to the
  bytewise variants); ICU is still required and linked unconditionally,
  so the whole project keeps building and both code paths remain
  testable.
- Cover the ICU-free comparators with new unit tests in
  `StringSortComparatorTest.cpp`.
